backend/requirements.txt
flask==3.0.3
pymongo==4.6.1
dnspython
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
gunicorn==21.2.0
xrpl-py==4.1.0